requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djoser
drf-yasg
